+++ src/main/java/br/com/rbernardes/familybudget/controller/IncomeController.java
@@ -1,7 +1,9 @@
 package br.com.rbernardes.familybudget.controller;
 
 import java.net.URI;
+import java.util.ArrayList;
 import java.util.List;
+import java.util.Optional;
 
 import javax.transaction.Transactional;
 import javax.validation.Valid;
@@ -38,15 +40,25 @@ public class IncomeController {
 	}
 	
 	@GetMapping(value = "/{id}")
-	public IncomeDTO expenseById(@PathVariable @Valid Long id) {
-		Income income = incomeRepository.getReferenceById(id);		
-		return new IncomeDTO(income);
+	public ResponseEntity<IncomeDTO> incomeById(@PathVariable @Valid Long id) {
+		Optional<Income> income = incomeRepository.findById(id);
+		if(income.isPresent()) {
+			return ResponseEntity.ok(new IncomeDTO(income.get()));
+		}
+		return ResponseEntity.notFound().build();
 	}
 	
 	@PostMapping
 	@Transactional
 	public ResponseEntity<IncomeDTO> addIncome(@RequestBody @Valid IncomeForm incomeForm, UriComponentsBuilder uriBuilder) {
 		Income income = incomeForm.convertToIncome();
+		List<Income> incomesList = new ArrayList<>();
+		incomesList = incomeRepository.findAll();
+		for(int i = 0; i < incomesList.size(); i++) {
+			if(incomesList.get(i).getDescription() == income.getDescription() && incomesList.get(i).getDate() == income.getDate()) {
+			return ResponseEntity.ok(new IncomeDTO(income));
+			}
+		}
 		incomeRepository.save(income);
 		
 		URI uri = uriBuilder.path("/incomes/{id}").buildAndExpand(income.getId()).toUri();
@@ -56,14 +68,23 @@ public class IncomeController {
 	@PutMapping(value = "/{id}")
 	@Transactional
 	public ResponseEntity<IncomeDTO> incomeUpdate(@PathVariable @Valid Long id, @RequestBody @Valid UpdateIncomeForm updateIncomeForm) {
-		Income income = updateIncomeForm.update(id, incomeRepository);
-		return ResponseEntity.ok(new IncomeDTO(income));
+		Optional<Income> income = incomeRepository.findById(id);
+		if(income.isPresent()) {
+			income = Optional.of(updateIncomeForm.update(id, incomeRepository));
+			return ResponseEntity.ok(new IncomeDTO(income.get()));
+		}
+		
+		return ResponseEntity.notFound().build();
 	}
 	
 	@DeleteMapping(value = "/{id}")
 	@Transactional
 	public ResponseEntity<?> incomeDelete(@PathVariable @Valid Long id) {
-		incomeRepository.deleteById(id);		
-		return ResponseEntity.ok().build();
+		Optional<Income> income = incomeRepository.findById(id);
+		if(income.isPresent()) {
+			incomeRepository.deleteById(id);		
+			return ResponseEntity.ok().build();
+		}
+		return ResponseEntity.notFound().build();
 	}
 }
